+++ packages/plugin-intl/src/index.ts
@@ -5,7 +5,18 @@ import type { Plugin } from '@ice/app/types';
 
 const _dirname = path.dirname(fileURLToPath(import.meta.url));
 
-const plugin: Plugin = () => ({
+interface PluginOptions {
+  // The key of locale content read from the window object.
+  localeMessagesKey?: string;
+  defaultLocaleKey?: string;
+  useCDN?: boolean;
+}
+
+const plugin: Plugin<PluginOptions> = ({
+  localeMessagesKey = '__LOCALE_MESSAGES__',
+  defaultLocaleKey = '__DEFAULT_LOCALE__',
+  useCDN = false,
+} = {}) => ({
   name: 'plugin-intl',
   setup: ({ generator, context, createLogger, watch }) => {
     const { rootDir } = context;
@@ -26,6 +37,8 @@ const plugin: Plugin = () => ({
         path.join(_dirname, '../templates/locales.ts.ejs'),
         'locales.ts',
         {
+          localeMessagesKey,
+          defaultLocaleKey,
           localeImport: locales.join('\n'),
           localeExport: localeExport.join('\n  '),
         },
@@ -34,7 +47,7 @@ const plugin: Plugin = () => ({
     const globRule = 'src/locales/*.{ts,js,json}';
     // Glob all locale files, and generate runtime options.
     const localeFiles = fg.sync(globRule, { cwd: rootDir });
-    if (localeFiles.length > 0) {
+    if (localeFiles.length > 0 && !useCDN) {
       // Filter the entry of locale files.
       const mainEntry = localeFiles.find((file) => file.match(/index\.(ts|js|json)$/));
       let runtimeSource = '';
@@ -59,7 +72,14 @@ const plugin: Plugin = () => ({
         }, 'both');
       }
     } else {
-      logger.warn('No locale files found, please check the `src/locales` folder.');
+      renderLocaleEntry([]);
+      generator.addEntryImportAhead({
+        source: './locales',
+      // @ts-ignore
+      }, 'both');
+      if (!useCDN) {
+        logger.warn('No locale files found, please check the `src/locales` folder.');
+      }
     }
 
     // Add intl export from ice.